feat: inclui tratamento de excecao

+++ src/main/java/com/gabrielsilva/dscommerce/services/ProductServices.java
@@ -3,7 +3,11 @@ package com.gabrielsilva.dscommerce.services;
 import com.gabrielsilva.dscommerce.dto.ProductDTO;
 import com.gabrielsilva.dscommerce.entities.Product;
 import com.gabrielsilva.dscommerce.repositories.ProductRepository;
+import com.gabrielsilva.dscommerce.services.exceptions.DatabaseException;
+import com.gabrielsilva.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -20,8 +24,8 @@ public class ProductServices {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
+        Product product = repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -46,19 +50,28 @@ public class ProductServices {
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
 
-        Product entity = repository.getReferenceById(id);
-        
-        copyDtoToEntity(dto, entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
 
-        entity = repository.save(entity);
-
-        return new ProductDTO(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
 
     }
 
     @Transactional
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
